created addEstimatedDiffrenceAverageOldDataTest file in here we try to cluster the samples into two classes without knowing the true lable.
It didnt work so far.
1.we tried to cluster by the time by the covariance of the original data.
2. we tried to do pca first on 16 channels and then cluster on the time, didnt work very well
then we tried to cluster on the covariance and we reached 80 and 60 precent

then we wanted to remove all the distant points from every cluster and we got tired. - we still need to check it

we didnt try to do PCA on the time(all we did is pca on channels)